Package structure: moved custom recording layout.
Part of #963.
