+++ src/androidTest/java/de/dennisguse/opentracks/settings/PreferencesUtilsTest.java
@@ -20,9 +20,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import de.dennisguse.opentracks.R;
-import de.dennisguse.opentracks.content.data.DataField;
-import de.dennisguse.opentracks.content.data.Layout;
 import de.dennisguse.opentracks.io.file.TrackFileFormat;
+import de.dennisguse.opentracks.ui.customRecordingLayout.DataField;
+import de.dennisguse.opentracks.ui.customRecordingLayout.Layout;
 
 @RunWith(AndroidJUnit4.class)
 public class PreferencesUtilsTest {
