+++ src/androidTest/java/de/dennisguse/opentracks/ui/customRecordingLayout/EspressoCustomLayoutTest.java
@@ -1,4 +1,4 @@
-package de.dennisguse.opentracks;
+package de.dennisguse.opentracks.ui.customRecordingLayout;
 
 
 import static androidx.test.espresso.Espresso.onView;
@@ -20,6 +20,7 @@ import androidx.test.espresso.ViewAction;
 import androidx.test.espresso.ViewAssertion;
 import androidx.test.espresso.ViewInteraction;
 import androidx.test.espresso.action.MotionEvents;
+import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
@@ -32,8 +33,8 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import de.dennisguse.opentracks.content.data.DataField;
-import de.dennisguse.opentracks.content.data.Layout;
+import de.dennisguse.opentracks.R;
+import de.dennisguse.opentracks.TrackListActivity;
 import de.dennisguse.opentracks.settings.PreferencesUtils;
 
 @LargeTest
@@ -49,7 +50,7 @@ public class EspressoCustomLayoutTest {
     @Test
     public void customLayoutTest() {
         // TrackListActivity: start recording
-        ViewInteraction trackControllerRecordButton = onView(withId(R.id.controller_record));
+        ViewInteraction trackControllerRecordButton = onView(ViewMatchers.withId(R.id.controller_record));
         trackControllerRecordButton.perform(click());
 
         // Get custom layout preferences and check all data fields are showed.
